Addressed a few compiler warnings: - bad-function-cast - conditional-uninitialized - documentation-unknown-command (*) - float-conversion - format-security - missing-noreturn - missing-prototypes - newline-eof - static-in-inline - unreachable-code-break
(*) Warnings addressed were related to @-commands in docblocks.

+++ src/fbmath.c
@@ -28,7 +28,7 @@
 double
 _int_f_rand(void)
 {
-    return ((double)rand() / (double)RAND_MAX);
+    return (rand() / (double)RAND_MAX);
 }
 
 /**
